Ignore errors in files that aren’t in the current project (#17)
In some cases it is possible for SBT to generate an error that refers
to a file that is not in the local project. E.g., this occurs if you
are using a testing library and the error is actually triggered in
the library rather than in your code. Previously SublimeSBT would
crash in this circumstance since it assumed that the errors always
referred to files in the local project.