Always show the selection position in the sbt output panel
Previously every time you displayed the output panel it would show the end of
the view. This change means that we always show the position around the
selection's location in the panel. In other words, you can move around in the
view (e.g., to scroll back to where a message of interest is displayed), put
the selection there and that position will be restored next time the view is
displayed. The previous behaviour is preserved if you don't do anything since
the selection will be positioned at the end of the view each time a new
command is executed.