Add a test failure parser for multi-line failures
To date we've supported test failures that are on a single line. I.e.
of this format:

    [error|info] ... (filename::nn)

Some scalatest tests can produce multiline failures in situations where the
expected or found values print as multiline strings. In this case the error
will look something like this:

    [info] - test description here *** FAILED ***
    [info] ...
    [info] ... (filename:nn)

This change adds a test failure parser for failures of this form so that the
plugin can correctly display the failure message and move to the failure
location as per the previious behaviour for single line failures.