Add support for global and project-specific history initialisation
Previously the command history was always initialised to an empty list.
This change implements a capability to set up useful commands and have
them used to initalise the command history. Commands are obtained from
the .SublimeSBT_history file at the top level of an SBT project (for
project-specific history) and from the SublimeSBT setting called "history"
(for commands that apply to all projects).